+++ src/civis/resources/_resources.py
@@ -3,6 +3,8 @@ from functools import lru_cache
 import json
 import os
 import re
+import sys
+import time
 import textwrap
 from inspect import Signature, Parameter
 
@@ -537,7 +539,6 @@ def parse_api_spec(api_spec, api_version):
     return classes
 
 
-@lru_cache(maxsize=4)
 def get_api_spec(api_key, api_version="1.0", user_agent="civis-python"):
     """Download the Civis API specification.
 
@@ -549,7 +550,7 @@ def get_api_spec(api_key, api_version="1.0", user_agent="civis-python"):
         The version of endpoints to call. May instantiate multiple client
         objects with different versions.  Currently only "1.0" is supported.
     user_agent : string, optional
-        Provide this user agent to the the Civis API, along with an
+        Provide this user agent to the Civis API, along with an
         API client version tag and ``requests`` version tag.
     """
     if api_version == "1.0":
@@ -568,7 +569,35 @@ def get_api_spec(api_key, api_version="1.0", user_agent="civis-python"):
     return spec
 
 
+def _running_interactively():
+    # https://stackoverflow.com/a/64523765
+    return hasattr(sys, "ps1")
+
+
+@lru_cache
+def _spec_expire_time():
+    """Return the duration (seconds) after which a cached API spec expires."""
+    if _running_interactively():
+        return 60 * 15  # 15 minutes
+    else:
+        return 60 * 60 * 24  # 24 hours
+
+
+def _get_ttl_hash():
+    """Return the same value within `seconds` time period."""
+    seconds = _spec_expire_time()
+    return round(time.time() / seconds)
+
+
 @lru_cache(maxsize=4)
+def generate_classes_ttl_cache(api_key, api_version, ttl_hash):
+    """Wraps generate_classes with a time-to-live cache.
+
+    https://stackoverflow.com/a/55900800
+    """
+    return generate_classes(api_key, api_version)
+
+
 def generate_classes(api_key, api_version="1.0"):
     """Dynamically create classes to interface with the Civis API.
 
@@ -617,10 +646,22 @@ def cache_api_spec(cache=CACHED_SPEC_PATH, api_key=None, api_version="1.0"):
         json.dump(spec, _fout)
 
 
-def generate_classes_maybe_cached(cache, api_key, api_version):
+def generate_classes_maybe_cached(
+    cache,
+    api_key,
+    api_version,
+    force_refresh_api_spec=False,
+):
     """Generate class objects either from /endpoints or a local cache."""
+    if cache and force_refresh_api_spec:
+        raise TypeError(
+            "you cannot provide a local spec and "
+            "clear the in-memory cached specs at the same time"
+        )
+    if force_refresh_api_spec:
+        generate_classes_ttl_cache.cache_clear()
     if cache is None:
-        classes = generate_classes(api_key, api_version)
+        classes = generate_classes_ttl_cache(api_key, api_version, _get_ttl_hash())
     else:
         if isinstance(cache, OrderedDict):
             raw_spec = cache