+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   which is the version where `joblib.parallel_config` was introduced and
   `joblib.parallel_backend` was deprecated. (#488)
 - Improved `import civis` startup time with a 5x speed boost. (#490)
+- Improved the startup time of `import civis` with a 5x speed boost. (#490)
+- The downloaded API spec due to the `civis.APIClient` instantiation is now
+  a time-to-live cache in memory (15 minutes for interactive Python, or 24 hours in scripts). (#491)
 
 ### Deprecated
 ### Removed